+++ examples/show-image-tch.rs
@@ -1,3 +1,4 @@
+use show_image::Event;
 use show_image::ImageData;
 use show_image::make_window;
 use show_image::tch::TensorAsImage;
@@ -22,8 +23,8 @@ fn main() -> Result<(), String> {
 	let window = make_window("image")?;
 	window.set_image(name, image)?;
 
-	while let Ok(event) = window.wait_key(std::time::Duration::from_millis(100)) {
-		if let Some(event) = event {
+	for event in window.events()? {
+		if let Event::KeyboardEvent(event) = event {
 			//println!("{:#?}", event);
 			if event.key == show_image::KeyCode::Escape {
 				break;