+++ CHANGELOG.md
@@ -3,6 +3,8 @@ v0.7.0
   * Add support for showing overlays on top of images.
   * Fold consecutive mouse move events to reduce number of events.
   * Allow registered event handlers to remove themselves.
+  * Remove implicit event channel of window handles.
+  * Make window handles Clone.
 
 v0.6.2
   * Add support for `raqote::DrawTarget` and `raqote::Image`.